backend/app/app.py
import os
import base64
import tempfile
import asyncio
from typing import Optional, Dict, Any
from fastapi import FastAPI, HTTPException, Request, Depends
from fastapi.middleware.cors import CORSMiddleware
from pydantic import BaseModel
import httpx
from bs4 import BeautifulSoup
from functools import lru_cache
from dotenv import load_dotenv
import numpy as np
import soundfile as sf
from faster_whisper import WhisperModel
from app.services.legal_service import LegalService

load_dotenv()

app = FastAPI(title="Legals MVP API")

app.add_middleware(
    CORSMiddleware,
    allow_origins=["http://localhost:3000", "http://localhost:3001"],
    allow_credentials=True,
    allow_methods=["*"],
    allow_headers=["*"],
)

SARVAM_API_KEY = os.getenv("SARVAM_API_KEY")
WHISPER_MODEL_SIZE = os.getenv("WHISPER_MODEL_SIZE", "base")

# --- Services ---
legal_service = LegalService()

# --- Models ---
class LegalQuery(BaseModel):
    text: str
    language: str = "hi"
    constituency: Optional[str] = None

class VoiceQuery(BaseModel):
    audio_base64: str
    session_id: str

# --- Utilities ---

@lru_cache(maxsize=1)
def get_whisper_model():
    # Use CPU by default for MVP, can be changed to 'cuda' if GPU available
    device = "cpu"
    return WhisperModel(WHISPER_MODEL_SIZE, device=device, compute_type="int8")

async def get_sarvam_llm_response(prompt: str, language: str = "hi"):
    if not SARVAM_API_KEY:
        return "Sarvam API Key not configured. (Mock response: Legal guidance received.)"
    
    url = "https://api.sarvam.ai/v1/chat/completions"
    headers = {"api-subscription-key": SARVAM_API_KEY}
    payload = {
        "model": "sarvam-30b",
        "messages": [
            {"role": "system", "content": f"You are a legal assistant specializing in Indian law and harassment cases. Respond in {language}."},
            {"role": "user", "content": prompt}
        ]
    }
    
    async with httpx.AsyncClient() as client:
        try:
            response = await client.post(url, json=payload, headers=headers, timeout=30.0)
            response.raise_for_status()
            return response.json()["choices"][0]["message"]["content"]
        except httpx.HTTPStatusError as e:
            return f"Error calling Sarvam AI: {e.response.text}"
        except Exception as e:
            return f"Error calling Sarvam AI: {str(e)}"

async def text_to_speech_sarvam(text: str, language_code: str = "hi-IN"):
    if not SARVAM_API_KEY:
        return None
    
    url = "https://api.sarvam.ai/text-to-speech"
    headers = {"api-subscription-key": SARVAM_API_KEY}
    payload = {
        "inputs": [text],
        "target_language_code": language_code,
        "speaker": "anushka", # Default speaker
        "pitch": 0,
        "pace": 1,
        "loudness": 1.5,
        "speech_sample_rate": 24000
    }
    
    async with httpx.AsyncClient() as client:
        try:
            response = await client.post(url, json=payload, headers=headers, timeout=30.0)
            response.raise_for_status()
            # Sarvam returns a list of base64 encoded audio strings or direct bytes?
            # Usually it's base64 in JSON for this endpoint format.
            return response.json()["audios"][0]
        except httpx.HTTPStatusError as e:
            print(f"TTS HTTP Error: {e.response.text}")
            return None
        except Exception as e:
            print(f"TTS Error: {e}")
            return None

# --- Scraping Logic ---

# Cache for constituency data
constituency_cache: Dict[str, Any] = {}

async def scrape_mplads_data(lookup_val: str):
    # This is a placeholder for actual scraping logic since the portal is complex
    # In a real scenario, we'd navigate the dashboard and extract tables
    # For MVP, we'll return mock data if scraping is too complex for a single tool call
    # or if we can't access the site.
    
    if lookup_val in constituency_cache:
        return constituency_cache[lookup_val]
    
    # Mock data for demonstration if scraping fails or is not implemented fully
    mock_data = {
        "constituency": "Kolkata South",
        "mp_name": "Mala Roy",
        "total_funds_allocated": "17.5 Cr",
        "funds_utilized": "14.2 Cr",
        "projects_completed": 124,
        "projects_ongoing": 12,
        "status": "Active"
    }
    
    # Simple check for PIN or Name
    if "700026" in lookup_val: # Example PIN for Kolkata South
        constituency_cache[lookup_val] = mock_data
        return mock_data
        
    # Attempting a simple GET to the dashboard to see if we can find anything
    try:
        url = "https://mplads.mospi.gov.in/mplads/Public_Dashboard/Dashboard.aspx"
        async with httpx.AsyncClient() as client:
            resp = await client.get(url, timeout=10.0)
            if resp.status_code == 200:
                # Actual parsing would go here
                pass
    except:
        pass
        
    return mock_data

# --- Endpoints ---

@app.post("/api/query")
async def general_query(query: LegalQuery):
    # Use the guardrailed local RAG pipeline
    result = await asyncio.to_thread(legal_service.generate_with_guardrails, query.text)
    return {
        "reply": result["answer"],
        "metadata": {
            "model": result["model_used"],
            "judge_verified": result["verified"],
            "feedback": result["judge_feedback"]
        }
    }

@app.get("/api/query/constituency/{mp_name_or_pin}")
async def mp_performance(mp_name_or_pin: str):
    data = await scrape_mplads_data(mp_name_or_pin)
    return data

@app.post("/api/chat/voice")
async def voice_chat(query: VoiceQuery):
    # 1. Decode audio
    try:
        audio_data = base64.b64decode(query.audio_base64)
        with tempfile.NamedTemporaryFile(suffix=".wav", delete=False) as tmp:
            tmp.write(audio_data)
            tmp_path = tmp.name
    except Exception as e:
        raise HTTPException(status_code=400, detail=f"Invalid audio data: {str(e)}")
    
    try:
        # 2. Transcribe
        model = get_whisper_model()
        segments, _ = await asyncio.to_thread(model.transcribe, tmp_path)
        transcription = " ".join([seg.text for seg in segments])
        
        # 3. Get LLM Reply
        reply_text = await get_sarvam_llm_response(transcription)
        
        # 4. TTS
        audio_reply_base64 = await text_to_speech_sarvam(reply_text)
        
        return {
            "transcription": transcription,
            "reply": reply_text,
            "audio_out": audio_reply_base64
        }
    except Exception as e:
        raise HTTPException(status_code=500, detail=f"Voice processing error: {str(e)}")
    finally:
        if os.path.exists(tmp_path):
            os.unlink(tmp_path)

@app.get("/hello-world")
def read_root():
    return {"message":"Hello World"}