backend/pyproject.toml
[project]
name = "legals"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastapi>=0.135.1",
    "uvicorn>=0.42.0",
    "httpx>=0.28.1",
    "beautifulsoup4>=4.13.3",
    "faster-whisper>=1.1.1",
    "numpy>=2.2.3",
    "python-dotenv>=1.0.1",
    "soundfile>=0.13.1",
    "pydantic-settings>=2.8.1",
    "python-multipart>=0.0.12",
    "ollama>=0.6.1",
    "pydantic>=2.12.5",
    "requests>=2.32.5",
    "tqdm>=4.67.3",
    "langchain-chroma>=0.2.1",
    "langchain-huggingface>=0.1.1",
    "protobuf<=3.20.3",
    "sentence-transformers>=3.0.1",
]
